+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.0] - 2025-06-15
+
 ### Added
 
 - Added new check `CheckHtmlWithTidy` https://github.com/StaticPipes/StaticPipes/issues/40